Sequential: Added malloc-less implementation
Instead of allocating new arrays for points and ids, the original array is used accross all iterations along with an array of ids that belong in the subtree. ids and distances are parallel-synced: while quickselect sorts the distances to left and right of the median , it also swaps elements of the ids array so a point with ids[i] (index in the original points array) has distance[i] from the vantage point.